requirements.txt
asgiref==3.4.1
astroid==2.9.2
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-crispy-forms==1.13.0
django-import-export==2.7.1
et-xmlfile==1.1.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
platformdirs==2.4.1
psycopg2==2.9.3
pylint==2.12.2
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
typing_extensions==4.0.1
tzdata==2021.5
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
